Move rainbow to its own component
You know it's bad when you start prefixing class names when you have a library that does that for you
