+++ src/components/store/Hero/style.module.scss.d.ts
@@ -1,25 +1,15 @@
 export type Styles = {
-  barsLeft: string;
-  barsRest: string;
-  barsRight: string;
   deco1: string;
   deco2: string;
   description: string;
   heading: string;
   hero: string;
   heroContent: string;
-  heroWrapper: string;
   imageWrapper: string;
   left: string;
-  orange: string;
-  pink: string;
-  rainbow: string;
-  rainbowContent: string;
-  rainbowWrapper: string;
-  red: string;
   right: string;
-  teal: string;
   textButton: string;
+  wrapper: string;
 };
 
 export type ClassNames = keyof Styles;
